rename db
