+++ services/web/server/src/simcore_service_webserver/director_v2_handlers.py
@@ -19,7 +19,7 @@ from .director_v2_core_computations import ComputationsApi
 from .director_v2_exceptions import DirectorServiceError
 from .login.decorators import RQT_USERID_KEY, login_required
 from .security_decorators import permission_required
-from .version_control.version_control_db import CommitID
+from .version_control.db import CommitID
 
 log = logging.getLogger(__name__)
