+++ services/web/server/tests/unit/with_dbs/03/version_control/test_version_control.py
@@ -9,10 +9,8 @@ from models_library.projects import Workbench
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from pydantic import BaseModel
 from simcore_service_webserver._meta import API_VTAG as VX
+from simcore_service_webserver.version_control.db import compute_workbench_checksum
 from simcore_service_webserver.version_control.plugin import _rest_handlers
-from simcore_service_webserver.version_control.version_control_db import (
-    compute_workbench_checksum,
-)
 
 ProjectDict = dict[str, Any]
